Drastically rewrite to show simpler server
The existing code contained lots of setup mixed with running, and was
wired for an in-process client only.  This made for a confusing sample,
plus one that was rather inflexible.  This setup splits things into cmd
pieces which do argument processing, a bootstrapper which generates a
more robust config, and finally a small server component which is quite
a bit more readable.

Signed-off-by: Jason Yellick <jyellick@us.ibm.com>